requirements.txt
beautifulsoup4==4.12.2
selenium==4.12.0
pyTelegramBotAPI==4.13.0
python-dotenv==1.0.0